+++ examples/streamz/python/dga_detection.py
@@ -12,130 +12,42 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import gc
-import sys
 import time
 import dask
-import torch
-import signal
-from streamz import Stream
-from tornado import ioloop
 from clx_streamz_tools import utils
-
-def inference(messages_df):
-    # Messages will be received and run through DGA inferencing
-    worker = dask.distributed.get_worker()
-    batch_start_time = int(round(time.time()))
-    result_size = messages_df.shape[0]
-    print("Processing batch size: " + str(result_size))
-    dd = worker.data["dga_detector"]
-    preds = dd.predict(messages_df["domain"])
-    messages_df["preds"] = preds
-    torch.cuda.empty_cache()
-    gc.collect()
-    return (messages_df, batch_start_time, result_size)
+from clx_streamz_tools import streamz_workflow
 
 
-def sink_to_kafka(processed_data):
-    # Prediction data will be published to provided kafka producer
-    messages_df = processed_data[0]
-    utils.kafka_sink(producer_conf, args.output_topic, messages_df)
-    return processed_data
+class DGADetectionWorkflow(streamz_workflow.StreamzWorkflow):
+    def inference(self, messages_df):
+        # Messages will be received and run through DGA inferencing
+        worker = dask.distributed.get_worker()
+        batch_start_time = int(round(time.time()))
+        result_size = messages_df.shape[0]
+        print("Processing batch size: " + str(result_size))
+        dd = worker.data["dga_detector"]
+        preds = dd.predict(messages_df["domain"])
+        messages_df["preds"] = preds
+        return (messages_df, batch_start_time, result_size)
 
+    def worker_init(self):
+        # Initialization for each dask worker
+        from clx.analytics.dga_detector import DGADetector
 
-def worker_init():
-    # Initialization for each dask worker
-    from clx.analytics.dga_detector import DGADetector
-
-    worker = dask.distributed.get_worker()
-    dd = DGADetector()
-    print(
-        "Initializing Dask worker: "
-        + str(worker)
-        + " with dga model. Model File: "
-        + str(args.model)
-    )
-    dd.load_model(args.model)
-    worker.data["dga_detector"] = dd
-    print("Successfully initialized dask worker " + str(worker))
-
-
-def signal_term_handler(signal, frame):
-    # Receives signal and calculates benchmark if indicated in argument
-    print("Exiting streamz script...")
-    if args.benchmark:
-        (time_diff, throughput_mbps, avg_batch_size) = utils.calc_benchmark(
-            output, args.benchmark
-        )
-        print("*** BENCHMARK ***")
+        worker = dask.distributed.get_worker()
+        dd = DGADetector()
         print(
-            "Job duration: {:.3f} secs, Throughput(mb/sec):{:.3f}, Avg. Batch size(mb):{:.3f}".format(
-                time_diff, throughput_mbps, avg_batch_size
-            )
+            "Initializing Dask worker: "
+            + str(worker)
+            + " with dga model. Model File: "
+            + str(self.args.model)
         )
-    sys.exit(0)
+        dd.load_model(self.args.model)
+        # this dict can be used for adding more objects to distributed dask worker
+        obj_dict = {"dga_detector": dd}
+        worker = utils.init_dask_workers(worker, self.config, obj_dict)
 
-
-def start_stream():
-    # Define the streaming pipeline.
-    # note: currently cudf engine supports only flatten json message format.
-    source = Stream.from_kafka_batched(
-        args.input_topic,
-        consumer_conf,
-        poll_interval=args.poll_interval,
-        # npartitions value varies based on kafka topic partitions configuration.
-        npartitions=1,
-        asynchronous=True,
-        dask=True,
-        engine="cudf",
-        max_batch_size=args.max_batch_size,
-    )
-    global output
-    # If benchmark arg is True, use streamz to compute benchmark
-    if args.benchmark:
-        print("Benchmark will be calculated")
-        output = (
-            source.map(inference)
-            .map(lambda x: (x[0], x[1], int(round(time.time())), x[2]))
-            .map(sink_to_kafka)
-            .gather()
-            .sink_to_list()
-        )
-    else:
-        output = source.map(inference).map(sink_to_kafka).gather()
-
-    source.start()
 
 if __name__ == "__main__":
-    # Parse arguments
-    args = utils.parse_arguments()
-    # Handle script exit
-    signal.signal(signal.SIGTERM, signal_term_handler)
-    signal.signal(signal.SIGINT, signal_term_handler)
-
-    client = utils.create_dask_client()
-    client.run(worker_init)
-
-    producer_conf = {
-        "bootstrap.servers": args.broker,
-        "session.timeout.ms": "10000",
-        # "queue.buffering.max.messages": "250000",
-        # "linger.ms": "100"
-    }
-    consumer_conf = {
-        "bootstrap.servers": args.broker,
-        "group.id": args.group_id,
-        "session.timeout.ms": "60000",
-        "enable.partition.eof": "true",
-        "auto.offset.reset": "earliest",
-    }
-    print("Producer conf: " + str(producer_conf))
-    print("Consumer conf: " + str(consumer_conf))
-    
-    loop = ioloop.IOLoop.current()
-    loop.add_callback(start_stream)
-    
-    try:
-        loop.start()
-    except KeyboardInterrupt:
-        loop.stop()
+    dga_detection = DGADetectionWorkflow()
+    dga_detection.start()