+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ## Improvements
 - PR #303 CLX sequence classifier
+- PR #308 CLX streamz workflow updates
+- PR #313 XGBoost training notebook
 
 ## Bug Fixes
 
-# clx 0.17.0 (Date TBD)
+# clx 0.17.0 (10 Dec 2020)
 
 ## New Features
 - PR #277 cyBERT class to support DistilBERT and ELECTRA